src: add a simple inference function to wit.rs
This function is a placeholder for the actual inference function that
will be used to call the wasi-nn inference function. This is currently
and issue when trying to transform a core wasm module to a component and
it may be the case that we need to create a component adapter for
wasi-nn.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>