src: add/stash working inference llama_cpp backend
This commit adds the working inference for the llama_cpp backend. The
code is just a prototype to see if this actually works and I'm going to
iterate on this. At the moment the llama_cpp backend is using
llm-chain-llama which also required some changes so that private structs
can be accesses which I'm not sure will be considered a good idea. But
an alternative would be just use a smaller subset of that code if they
are alright with that to suite the needs of wasmtime better.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>