src: add generator component
This commit add a generator which is responsible for generating a new
component config module for a specified configuration. Currently the
configuration only consists of a model_path and a prompt and these are
taken as command line arguments. There will be a need for more
configuration parameters if this is something that we want to pursue in
which case it might be better to have a configuration file that is
passed in, perhaps in json format and then have a JSON object that can
deserialize it into a struct.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>