src: add initial binaryen generator
This is an investigation into using the binaryen crate to generate
wasm modules instead of the hidious solution that is currently in place
where we are creating a Cargo workspace and compiling the wasm module.

The motivation for this that if we can generate the wasm module that
would remove the requirement of having a rust toolchain installed on the
machine where the inf-wasi component is being built.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>