src: revert to older version of wasi-nn.wit
This commit reverts to an older version of the wasi-nn spec as this is
what the wasmtime version of wasi-nn is using. And this is what I've
updated the wasi-nn bindings to use as well. Hopefully this will get
updated.

This also means that we need to wait until wasmtime 20.0.0 is released.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>